Fix double type inference for math.sqrt and math.pow results
Agent-Logs-Url: https://github.com/snehadpatel/CompilerDesign/sessions/f9d39f1b-7030-441b-8f61-ca2995d71793

+++ outputs/math_output.c
@@ -5,11 +5,11 @@
 int main() {
     printf("Testing Math Library\n");
     float val = 16.0;
-    int root = sqrt(val);
+    double root = sqrt(val);
     printf("Square root of 16 is:\n");
-    printf("%d\n", root);
-    int p = pow(2.0, 3.0);
+    printf("%g\n", root);
+    double p = pow(2.0, 3.0);
     printf("2 to the power 3 is:\n");
-    printf("%d\n", p);
+    printf("%g\n", p);
     return 0;
 }
